Idempotency for JDK 7

+++ provisioning/ubuntudev/bootstrap.sh
@@ -35,10 +35,12 @@ function install_oracle_jdk_6 {
 }
 
 function install_oracle_jdk_7 {
-	wget --output-document=/tmp/vagrant/jdk-7u45-linux-x64.gz https://jacderida-software.s3.amazonaws.com/jdk-7u45-linux-x64.gz
-	tar -xvf /tmp/vagrant/jdk-7u45-linux-x64.gz
-	mv ./jdk1.7.0_45 ./java-7-oracle/
-	mv ./java-7-oracle /usr/lib/jvm/
+	if [ ! -e /usr/lib/jvm/java-7-oracle/bin/java ]; then
+		wget --output-document=/tmp/vagrant/jdk-7u45-linux-x64.gz https://jacderida-software.s3.amazonaws.com/jdk-7u45-linux-x64.gz
+		tar -xvf /tmp/vagrant/jdk-7u45-linux-x64.gz
+		mv ./jdk1.7.0_45 ./java-7-oracle/
+		mv ./java-7-oracle /usr/lib/jvm/
+	fi
 }
 
 function install_virtualbox {
